// src/app/models/setting/ORMSaveSetupLocation.ts
export class ORMSaveSetupLocation {
	location_id: string;
	practice_id: string;
	name: string;
	zip: string;
	city: string;
	state: string;
	address: string;
	address2: string;
	phone: string;
	fax: string;
	clia_number: string;
	clia_expiry_date: string;
	is_block: boolean;
	deleted: boolean;
	created_user: string;
	modified_user: string;
	client_date_created: string;
	client_date_modified: string;
	date_created: string;
	date_modified: string;
	facility_id: string;
}